moved channels and messages to separate file
